menampilkan jadwal dan kegiatan di home

+++ routes/web.php
@@ -4,7 +4,7 @@ use App\Models\User;
 use Illuminate\Support\Facades\Route;
 
 use App\Http\Controllers\PostController;
-
+use App\Http\Controllers\HomeController;
 use App\Http\Controllers\LoginController;
 use App\Http\Controllers\BannerController;
 use App\Http\Controllers\JadwalController;
@@ -12,12 +12,6 @@ use App\Http\Controllers\KontakController;
 use App\Http\Controllers\KegiatanController;
 use App\Http\Controllers\RegisterController;
 use App\Http\Controllers\DashboardPostController;
-
-Route::get('/', function () {
-    return view('home', [
-        "title" => "Home"
-    ]);
-});
 
 Route::get('/tentang', function () {
     return view('tentang', [
@@ -58,18 +52,18 @@ Route::get('/authors/{author:username}', function(User $author){
 
 Route::get('/dashboard' , function(){
     return view('dashboard.index');
-})->middleware('auth');
+})->middleware('admin');
 
-Route::get('/dashboard/posts/checkSlug', [DashboardPostController::class, 'checkSlug'])->middleware('auth');
+Route::get('/dashboard/posts/checkSlug', [DashboardPostController::class, 'checkSlug'])->middleware('auth', 'admin');
 Route::resource('/dashboard/posts', DashboardPostController::class)->middleware('auth', 'admin');
 
-Route::get('/dashboard/kontak', [KontakController::class, 'index']);
-Route::post('/dashboard/kontak', [KontakController::class, 'store']);
-Route::delete('/dashboard/kontak/{kontak}', [KontakController::class, 'destroy'])->name('kontak.destroy');
+Route::get('/dashboard/kontak', [KontakController::class, 'index'])->middleware('admin');
+Route::post('/dashboard/kontak', [KontakController::class, 'store'])->middleware('admin');
+Route::delete('/dashboard/kontak/{kontak}', [KontakController::class, 'destroy'])->name('kontak.destroy')->middleware('admin');
 
-Route::resource('dashboard/banner', BannerController::class);
-Route::resource('dashboard/jadwal', JadwalController::class);
-Route::resource('dashboard/kegiatan', KegiatanController::class);
+Route::resource('dashboard/banner', BannerController::class)->middleware('admin');
+Route::resource('dashboard/jadwal', JadwalController::class)->middleware('admin');
+Route::resource('dashboard/kegiatan', KegiatanController::class)->middleware('admin');
 
+Route::get('/', [HomeController::class, 'index']);
 
-Route::get('/', [BannerController::class, 'showBanner']);
